Refactor: Plant 상세 위젯 분리 #101
